Exit Code Testing
This change adds some testing around the status code used when os.Exit is
called.  Previously, intercepting this invocation was deemed too difficult.
However, as other buildpacks started development, it became clear that testing
this would be important and therefore determining a strategy for doing it, and
then applying that strategy to this library would be necessary.