Environment Variable existence
Previously, the only way to determine if an environment variable had been set
was to enumerate the entire collection and determine if any Name matched.
Given that this is likely to be a common behavior, this change adds a method
to determine if an Environment Variable of a given name is set.

Signed-off-by: Ben Hale <bhale@pivotal.io>