Debug Logging
Previously, setting BP_DEBUG to enable debug logging only worked outside of a
legal CNB runtime.  There was a race that the logger had to be created before
the platform could be created and used to load the environment.  This change
breaks the abstraction slightly and makes the Logger aware of where it would
need to look for this one particular environment variable so that it can be
read before the platform is created.

Signed-off-by: Ben Hale <bhale@pivotal.io>